+++ client/rpc_codec.go
@@ -4,8 +4,8 @@ import (
 	"bytes"
 
 	"github.com/micro/go-micro/codec"
-	"github.com/micro/go-micro/codec/json"
-	"github.com/micro/go-micro/codec/proto"
+	"github.com/micro/go-micro/codec/jsonrpc"
+	"github.com/micro/go-micro/codec/protorpc"
 	"github.com/micro/go-micro/transport"
 	rpc "github.com/youtube/vitess/go/rpcplus"
 )
@@ -27,11 +27,11 @@ var (
 	defaultContentType = "application/octet-stream"
 
 	defaultCodecs = map[string]codec.NewCodec{
-		"application/json":         json.NewCodec,
-		"application/json-rpc":     json.NewCodec,
-		"application/protobuf":     proto.NewCodec,
-		"application/proto-rpc":    proto.NewCodec,
-		"application/octet-stream": proto.NewCodec,
+		"application/json":         jsonrpc.NewCodec,
+		"application/json-rpc":     jsonrpc.NewCodec,
+		"application/protobuf":     protorpc.NewCodec,
+		"application/proto-rpc":    protorpc.NewCodec,
+		"application/octet-stream": protorpc.NewCodec,
 	}
 )
